# docs/swagger.yaml
basePath: /
definitions:
  entity.Account:
    properties:
      account_number:
        maxLength: 20
        minLength: 20
        type: string
      bank_id_number:
        maxLength: 9
        minLength: 9
        type: string
      bank_name:
        type: string
      id:
        type: integer
      organization_id:
        type: integer
    required:
    - account_number
    - bank_id_number
    - bank_name
    type: object
...
  entity.Mechanic:
    properties:
      full_name:
        type: string
      id:
        type: integer
    required:
    - full_name
    type: object
  entity.Organization:
    properties:
      address:
        type: string
      chief:
        type: string
      financial_chief:
        type: string
      id:
        type: integer
      inn_kpp:
        maxLength: 20
        minLength: 20
        type: string
      name:
        type: string
    required:
    - address
    - chief
    - financial_chief
    - inn_kpp
    - name
    type: object
...
  entity.UpdateAccountInput:
    properties:
      account_number:
        type: string
      bank_id_number:
        type: string
      bank_name:
        type: string
    type: object
...
  entity.UpdateMechanicInput:
    properties:
      full_name:
        type: string
    type: object
...
  handler.createAccountResponse:
    properties:
      account:
        $ref: '#/definitions/entity.Account'
    type: object
...
  handler.createMechanicResponse:
    properties:
      mechanic:
        $ref: '#/definitions/entity.Mechanic'
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.getAccountsResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/entity.Account'
        type: array
      organization:
        $ref: '#/definitions/entity.Organization'
    type: object
...
  handler.getMechanicsResponse:
    properties:
      mechanics:
        items:
          $ref: '#/definitions/entity.Mechanic'
        type: array
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
  handler.updateAccountResponse:
    properties:
      account:
        $ref: '#/definitions/entity.Account'
    type: object
...
  handler.updateMechanicResponse:
    properties:
      update_mechanic:
        $ref: '#/definitions/entity.Mechanic'
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API Server for Document Automation Service (Putlist)
  title: My Test App API
  version: "1.0"
paths:
...
  /{organization_id}/accounts:
    get:
      consumes:
      - application/json
      description: gets data about all bank accounts
      operationId: get-accounts
      parameters:
      - description: organization id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: get accounts
          schema:
            $ref: '#/definitions/handler.getAccountsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: creates new data about the bank account
      operationId: create-bank-account
      parameters:
      - description: organization id
        in: path
        name: organization_id
        required: true
        type: integer
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Account'
      produces:
      - application/json
      responses:
        "200":
          description: created account
          schema:
            $ref: '#/definitions/handler.createAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Account
      tags:
      - accounts
  /{organization_id}/accounts/{account_id}:
    delete:
      consumes:
      - application/json
      description: deletes bank account
      operationId: delete-account
      parameters:
      - description: organization id
        in: path
        name: organization_id
        required: true
        type: integer
      - description: account id
        in: path
        name: account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Account
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: updates data about the bank account
      operationId: update-account
      parameters:
      - description: organization id
        in: path
        name: organization_id
        required: true
        type: integer
      - description: account id
        in: path
        name: account_id
        required: true
        type: integer
      - description: update account info
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated account
          schema:
            $ref: '#/definitions/handler.updateAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Account
      tags:
      - accounts
...
  /mehanics:
    get:
      consumes:
      - application/json
      description: gets data about all mehanics
      operationId: get-mehanics
      produces:
      - application/json
      responses:
        "200":
          description: mehanics info
          schema:
            $ref: '#/definitions/handler.getMechanicsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Mehanics
      tags:
      - mehanics
    post:
      consumes:
      - application/json
      description: creates and adds data about the mehanic
      operationId: create-mehanic
      parameters:
      - description: mehanic info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Mechanic'
      produces:
      - application/json
      responses:
        "200":
          description: new mehanic
          schema:
            $ref: '#/definitions/handler.createMechanicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Mehanic
      tags:
      - mehanics
  /mehanics/{mehanic_id}:
    delete:
      consumes:
      - application/json
      description: deletes data about mehanic
      operationId: delete-mehanic
      parameters:
      - description: mehanic id
        in: body
        name: mehanic_id
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Mehanic
      tags:
      - mehanics
    put:
      consumes:
      - application/json
      description: updates data about the mehanic
      operationId: update-mehanic
      parameters:
      - description: mehanic id
        in: body
        name: mehanic_id
        required: true
        schema:
          type: integer
      - description: update mehanic info
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateMechanicInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated mehanic
          schema:
            $ref: '#/definitions/handler.updateMechanicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Mehanic
      tags:
      - mehanics
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"